requirements.txt
flask>=2.3.0
google-generativeai>=0.8.0
gunicorn>=21.2.0
